Move particle structs to particle files
