+++ src/openrct2/world/EntityTweener.h
@@ -7,7 +7,10 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
+#pragma once
+
 #include "SpriteBase.h"
+
 #include <vector>
 
 class EntityTweener
