+++ src/openrct2/world/Sprite.cpp
@@ -445,88 +445,6 @@ SpriteBase* CreateEntityAt(const uint16_t index, const EntityType type)
     PrepareNewEntity(entity, type);
     return entity;
 }
-/**
- *
- *  rct2: 0x00673200
- */
-void SteamParticle::Update()
-{
-    // Move up 1 z every 3 ticks (Starts after 4 ticks)
-    Invalidate();
-    time_to_move++;
-    if (time_to_move >= 4)
-    {
-        time_to_move = 1;
-        MoveTo({ x, y, z + 1 });
-    }
-    frame += 64;
-    if (frame >= (56 * 64))
-    {
-        sprite_remove(this);
-    }
-}
-
-/**
- *
- *  rct2: 0x0067363D
- */
-void sprite_misc_explosion_cloud_create(const CoordsXYZ& cloudPos)
-{
-    auto* sprite = CreateEntity<ExplosionCloud>();
-    if (sprite != nullptr)
-    {
-        sprite->sprite_width = 44;
-        sprite->sprite_height_negative = 32;
-        sprite->sprite_height_positive = 34;
-        sprite->MoveTo(cloudPos + CoordsXYZ{ 0, 0, 4 });
-        sprite->frame = 0;
-    }
-}
-
-/**
- *
- *  rct2: 0x00673385
- */
-void ExplosionCloud::Update()
-{
-    Invalidate();
-    frame += 128;
-    if (frame >= (36 * 128))
-    {
-        sprite_remove(this);
-    }
-}
-
-/**
- *
- *  rct2: 0x0067366B
- */
-void sprite_misc_explosion_flare_create(const CoordsXYZ& flarePos)
-{
-    MiscEntity* sprite = CreateEntity<ExplosionFlare>();
-    if (sprite != nullptr)
-    {
-        sprite->sprite_width = 25;
-        sprite->sprite_height_negative = 85;
-        sprite->sprite_height_positive = 8;
-        sprite->MoveTo(flarePos + CoordsXYZ{ 0, 0, 4 });
-        sprite->frame = 0;
-    }
-}
-
-/**
- *
- *  rct2: 0x006733B4
- */
-void ExplosionFlare::Update()
-{
-    Invalidate();
-    frame += 64;
-    if (frame >= (124 * 64))
-    {
-        sprite_remove(this);
-    }
-}
 
 template<typename T> void MiscUpdateAllType()
 {
